rename controller virtualworkspaceurls to synctarget
