+++ pkg/reconciler/workload/synctarget/synctarget_reconcile_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package virtualworkspaceurls
+package synctarget
 
 import (
 	"reflect"
